Add minor adjustments + link  redirections
