+++ src/index.js
@@ -1,7 +1,6 @@
 import NavFooter from "./assets/js/navAndFooter";
-import Home from "./assets/js/home";
-import Service from "./assets/js/service";
-import Visit from "./assets/js/visit-us";
+import Restaurant from "./assets/js/restaurant";
+
 import "./assets/bootstrap/css/bootstrap.min.css";
 const importAll = (r) => {
   return r.keys().map(r);
@@ -24,6 +23,7 @@ const contentChilds = () => {
   const tabs = document.createElement("div");
   const footer = document.createElement("div");
   nav.setAttribute("id", "nav");
+  nav.classList.add("sticky-top");
   header.setAttribute("id", "header");
   tabs.setAttribute("id", "tab");
   footer.setAttribute("id", "footer");
@@ -35,15 +35,15 @@ const contentChilds = () => {
 
 const renderTabs = (() => {
   const homeTab = () => {
-    Home.renderHome(images[1]);
+    Restaurant.renderHome(images[1]);
   };
 
   const serviceTab = () => {
-    Service.renderService(images[2], images[3], images[4]);
+    Restaurant.renderService(images[2], images[3], images[4]);
   };
 
   const visitTab = () => {
-    Visit.renderVisit();
+    Restaurant.renderVisit();
   };
   return { homeTab, serviceTab, visitTab };
 })();
@@ -54,12 +54,14 @@ document.addEventListener("DOMContentLoaded", () => {
   NavFooter.renderNav();
   NavFooter.renderHeader();
   NavFooter.renderFooter();
-  Home.renderHome(images[1]);
+  Restaurant.renderHome(images[1]);
   const homeItem = document.getElementById("home");
   const serviceItem = document.getElementById("service");
   const visitItem = document.getElementById("visit-us");
+  const logo = document.getElementById("logo");
 
   homeItem.addEventListener("click", renderTabs.homeTab);
   serviceItem.addEventListener("click", renderTabs.serviceTab);
   visitItem.addEventListener("click", renderTabs.visitTab);
+  logo.addEventListener("click", renderTabs.homeTab);
 });
